Move category page components to separate directory
