+++ saleor/static/js/components/categoryPage/AttributeInput.js
