+++ saleor/static/js/components/categoryPage/CategoryFilter.js
