+++ saleor/static/js/components/categoryPage/CategoryPage.js
