+++ saleor/static/js/components/categoryPage/PriceFilter.js
